+++ crates/thyllore-vulkan-core/src/renderer/mod.rs
@@ -1,9 +1,26 @@
 pub mod auto_exposure;
 pub mod bloom;
+pub mod composite;
 pub mod dof;
+pub mod gbuffer;
+pub mod line_mesh_draw;
+pub mod onion_skin;
+pub mod onion_skin_buffers;
+pub mod push_constants;
+pub mod rayquery;
 pub mod tonemap;
 
 pub use auto_exposure::record_auto_exposure_pass;
 pub use bloom::record_bloom_pass;
+pub use composite::{
+    begin_composite_render_pass, end_composite_render_pass, record_composite_draw,
+    record_composite_to_hdr_pass,
+};
 pub use dof::record_dof_pass;
+pub use gbuffer::record_gbuffer_pass;
+pub use line_mesh_draw::{record_line_mesh_draw, LineMeshDrawOptions};
+pub use onion_skin::{record_onion_skin_composite_pass, record_onion_skin_ghost_pass};
+pub use onion_skin_buffers::{OnionSkinGhostBuffer, OnionSkinGpuState};
+pub use push_constants::{GBufferPushConstants, OnionSkinPushConstants};
+pub use rayquery::record_ray_query_pass;
 pub use tonemap::{begin_tonemap_render_pass, end_tonemap_render_pass, record_tonemap_draw};
