+++ src/ecs/systems/phases/mod.rs
@@ -14,6 +14,7 @@ mod dispatch_timeline;
 pub mod event_dispatch_phase;
 pub mod input_phase;
 pub mod onion_skin_phase;
+pub mod render_phase;
 pub mod render_prep_phase;
 pub mod transform_phase;
 
@@ -21,5 +22,6 @@ pub use animation_phase::*;
 pub use event_dispatch_phase::*;
 pub use input_phase::*;
 pub use onion_skin_phase::*;
+pub use render_phase::build_frame_render_context;
 pub use render_prep_phase::*;
 pub use transform_phase::*;
