+++ src/vulkanr/renderer/deferred/mod.rs
@@ -1,13 +1,7 @@
-mod composite;
 pub mod gbuffer;
-mod onion_skin;
 mod overlay_renderer;
 mod pass_recording;
-mod rayquery;
 
-pub use composite::CompositePass;
-pub use gbuffer::{create_gbuffer_framebuffer, GBufferPass};
-pub use onion_skin::OnionSkinRenderPass;
+pub use gbuffer::create_gbuffer_framebuffer;
 pub use overlay_renderer::OverlayRenderer;
 pub use pass_recording::*;
-pub use rayquery::RayQueryPass;
